+++ searcher/model/item.go
@@ -2,5 +2,5 @@ package model
 
 type SliceItem struct {
 	Id    uint32
-	Score float32
+	Score int
 }
